integration: fix resultCount types to not overflow.
Using `uint8` as the type for the count fields in the `resultCount` type
produces overflows if more than 255 certificates with a given result
level are linted.

Our integration test corpus is just shy of 600,000 certificates so
`uint32` should be more than sufficient.

+++ v3/integration/result.go
@@ -9,10 +9,10 @@ import (
 )
 
 type resultCount struct {
-	FatalCount  uint8 `json:",omitempty"`
-	ErrCount    uint8 `json:",omitempty"`
-	WarnCount   uint8 `json:",omitempty"`
-	NoticeCount uint8 `json:",omitempty"`
+	FatalCount  uint32 `json:",omitempty"`
+	ErrCount    uint32 `json:",omitempty"`
+	WarnCount   uint32 `json:",omitempty"`
+	NoticeCount uint32 `json:",omitempty"`
 }
 
 // TODO(@cpu): Accept a threshold argument so that (for e.g. notices could be
